Env var update for each recipe

+++ recipes/genomics/Homo_sapiens/hg19/hg19-reference-genome-ucsc-v1/post-link.sh
@@ -23,18 +23,6 @@ fi
 
 mkdir -p $RECIPE_DIR
 
-recipe_env_name="ggd_hg19-reference-genome-ucsc-v1"
-recipe_env_name="$(echo "$recipe_env_name" | sed 's/-/_/g')"
-
-activate_dir="$env_dir/etc/conda/activate.d"
-deactivate_dir="$env_dir/etc/conda/deactivate.d"
-
-mkdir -p $activate_dir
-mkdir -p $deactivate_dir
-
-echo "export $recipe_env_name=$RECIPE_DIR" >> $activate_dir/env_vars.sh
-echo "unset $recipe_env_name">> $deactivate_dir/env_vars.sh
-
 (cd $RECIPE_DIR && bash $PKG_DIR/info/recipe/recipe.sh)
 
 cd $RECIPE_DIR
@@ -43,7 +31,51 @@ cd $RECIPE_DIR
 for f in *; do
     ext="${f#*.}"
     filename="{f%%.*}"
-    (mv $f "hg19-reference-genome-ucsc-v1.$ext")
+    if [[ ! -f "hg19-reference-genome-ucsc-v1.$ext" ]]  
+    then
+        (mv $f "hg19-reference-genome-ucsc-v1.$ext")
+    fi  
 done
+
+## Add environment variables 
+#### File
+if [[ `find $RECIPE_DIR -type f -maxdepth 1 | wc -l | sed 's/ //g'` == 1 ]] ## If only one file
+then
+    recipe_env_file_name="ggd_hg19-reference-genome-ucsc-v1_file"
+    recipe_env_file_name="$(echo "$recipe_env_file_name" | sed 's/-/_/g')"
+    file_path="$(find $RECIPE_DIR -type f -maxdepth 1)"
+
+elif [[ `find $RECIPE_DIR -type f -maxdepth 1 | wc -l | sed 's/ //g'` == 2 ]] ## If two files
+then
+    indexed_file=`find $RECIPE_DIR -type f \( -name "*.tbi" -or -name "*.fai" -or -name "*.bai" -or -name "*.crai" -or -name "*.gzi" \) -maxdepth 1`
+    if [[ ! -z "$indexed_file" ]] ## If index file exists
+    then
+        recipe_env_file_name="ggd_hg19-reference-genome-ucsc-v1_file"
+        recipe_env_file_name="$(echo "$recipe_env_file_name" | sed 's/-/_/g')"
+        file_path="$(echo $indexed_file | sed 's/\.[^.]*$//')" ## remove index extension
+    fi
+fi 
+
+#### Dir
+recipe_env_dir_name="ggd_hg19-reference-genome-ucsc-v1_dir"
+recipe_env_dir_name="$(echo "$recipe_env_dir_name" | sed 's/-/_/g')"
+
+activate_dir="$env_dir/etc/conda/activate.d"
+deactivate_dir="$env_dir/etc/conda/deactivate.d"
+
+mkdir -p $activate_dir
+mkdir -p $deactivate_dir
+
+echo "export $recipe_env_dir_name=$RECIPE_DIR" >> $activate_dir/env_vars.sh
+echo "unset $recipe_env_dir_name">> $deactivate_dir/env_vars.sh
+
+#### File
+    ## If the file env variable exists, set the env file var
+if [[ ! -z "${recipe_env_file_name:-}" ]] 
+then
+    echo "export $recipe_env_file_name=$file_path" >> $activate_dir/env_vars.sh
+    echo "unset $recipe_env_file_name">> $deactivate_dir/env_vars.sh
+fi
+    
 
 echo 'Recipe successfully built!'
